add rake task to process .js.erb files
To use the service worker we will need to cache assets in JavaScript
outside of the standard way Rails loads things. The primary use case (at
the moment) is when we cache assets in serviceworker.js. For example
when we want to cache "application.js" we need to know which
version (fingerprint) is current, but with esbuild we don't have access
to ERB and the `asset_path` helper that sorts this out for us.

This commit adds an ERB processor to .js files so we can use
`asset_path` and know the exact URL to cache in our service worker.