add public/serviceworker.js
This is a version of the ServiceWorker that caches pages and works when
"offline". It's a prototype to show the basics work.

This version isn't in ERB, as in previous commits, as it's unclear
whether we'll actually need to compile the current version of assets
into serviceworker.js ... these may be just get cached automatically
depending on how we implement it.