Log serving of static assets
Logging the serving of static assets isn't so useful, but while working
with service workers it can be particularly handy to have some
visibility on these. Especially if we plan to break out the JS used by
the service worker into separate files.

I couldn't find a way to log these through existing methods, so ...
pseudo-monkey-patching to the rescue.