Add esbuild task for serviceworker
The default asset pipeline for Rails makes all assets, including JS,
accessible on the site path '/assets/*'. However, the service worker
wants to (needs to?) live at the root of the site as the browser uses
it's location as the scope of the pages it controls. The specifics are a
bit murky, but it appears at laest that putting it at the site root is
current best practice.

To achieve this, and instead of moving all our site assets to the root
path, this change adds a separate esbuild task configured to just bulid
the service worker and place it into the public directory, where Rails
will serve it out as a static asset from the site root.

We may have to do further work to configure this for deployed
environments, we'll have to sort those out as we go.