Add cypress test framework
Cypress seems to be the best suited BDD framework for our purposes. Jest
was also considered, and I tried to add it, but ran into problems with
with transpiling code, and was abandoned, esp since for this spike a BDD
makes more sense.

I considered putting cypress into it's own repository, but I like
keeping the tests in lock-step with the version of the code they are
testing. Sticking it into this repo seems to be the simplest way forward
and I believe we can move it into a different repo later if we decide
that approach is better.

Cypress has been configured for e2e testing.