Add test for toggling offline and online.
Simple proof-of-concept test to test the online/offline switching
functionality. This DOES NOT fully test the functionality, however, as
only tests what's displayed on the page without testing that the browser
actually behaves as offline. That exercise has been punted down the road
for now.

This change also includes the decision to use `data-testid` property to
select page elements for use in tests, and adds a Cypress command to get
these easily in tests.