Import example campaign data.
This example data has been taken from the prototype and is used here to
populate our dev env db with useful data.

We include but don't use "schools_sample.json" yet, for now we only
focus on importing what's needed for the campaign in
"example-campaign.json". We can look to import the full schools list
later.