Replace seedbank with rake task
I started to refactor the seeds to be able to load different campaign
examples to be used by testing, but found that the whole seedbank
approach was being a pain and not well suited to loading a single
campaign example file, which is what we've got at the moment.

We may want to try to refactor this in the future, but for now this rake
task better suits the example data we've got to work with.