Rename children controller to vaccinations
I think children and vaccinations will be slightly different resources,
so I've removed the children controller and views to replace them with
vaccinations.  This includes:

- cleaned up the list of children for a campaign to look like the
  prototype
- added a dumbed-down journey to record a vaccination for a child,
  missing some steps that are in the prototype but still functional

As above, one of the decisions made here is that the list of children
for a campaign is actually a list of "vaccinations", and introduces a
controller to handle vaccinations specifically. It still calls them
"children" in the URL, however, as technically these are still children
resources belonging to the campaign. Breaking them out into their own
"vaccinations" controller helps group together the vaccination logic,
and pages, and is tidier than making the "children" controller modal,
which would require it to detect whether it is viewing children as a
part of a campaign or not, etc.