Build FHIR record for immunization
First step towards being able to send vaccine records to an external
FHIR server.

The FHIR builder here is simplistic and achieves the stepping-stone goal
of replicating the JSON that comes out of the POC POC, which is our
reference. As we have library support for FHIR we try to implement some
things there, but will need to continue adoption of the library's
conventions.

We also currently use many hard-coded values which will have to come
from the vaccination record.