Use FHIR::Dosage object
Aside from FHIR::Patient this appears to be the last object type that
was being represented by a hash when it should be a FHIR object. This
doesn't change the tests, the resulting JSON is identical, but the FHIR
object should hopefully add a level of validation and help with the
upgrade path should we need it.