Add vaccination history
This is not in the prototype yet but we've got a demo coming up and it
may help to demonstrate that this is real data going into a FHIR server.

This view is also horribly implemented, with a focus given solely to
having a commit ready ASAP. It is missing tests, proper abstraction, a
proper route, etc etc ... all to be fixed later once the prototype is
updated to have (or not have) a vaccination history.