Use window.location + ".json" for child JSON URL
Before this change, we relied on Ruby to generate the URL to retrieve
the JSON version of the child vaccination record. Now, with this change,
we just use our current location, and append ".json" to it.

This change builds up towards being able to emit a template version of
the child vaccine's show page, i.e. a version with no @child set, which
will be filled in via Javascript in the browser. This is a key step
towards enabling remote work.

This change also stands on it's own as changing to request the JSON
version of a record should simply be a matter of adding ".json" to the
path of the record.