Render child vaccination show page wihout @child
This change will allow us to render an empty show page for a child
vaccination, which will be retrieved by the client as a template to be
filled out with Javascript. The client can cache this template and
combine it with cached child vaccination JSON data to provide offline
capabilities.