Replace child vaccination template when offline
When running in offline mode, when a fetch request for a child
vaccination page is received, return the template page with no child
data filled in. The already-existing Stimulus controller on the child
vaccination page will fill in the child information from the cached JSON
version of the child data.