Add offline mode to request handlers
This allows the request handlers to decide how to respond based on
whether we're in offline mode or not. Offline mode is meant ensure we go
straight to cached responses, in the future this is likely to be done to
avoid network timeouts but for the moment, the same mechanism can be
used to simulate offline mode.