Add debugging to connectivity controller
For some reason, in our deployed environment the child vaccs show page
does not ever appear to get a response to the GET_CONNECTION_STATUS it
sends to the service worker. In fact, the service worker doesn't appear
to get the message even. This issue doesn't show up in local
envirionments, so this is being added temporarily to try to debug the
issue.