Replace cacheResponse with put
The cacheResponse function always seemed a bit weird, that you basically
give it a `response` object and all it does is immediately return that
response, while also doing something async with another parameter. The
fact that you have to `response.clone()` the return value is
unintuitive. It's easier to shrink it down to a thin wrapper on the
`put` function from the cache API, and then separate the cache/return
steps in the respective call sites.