Configure Jest to work with our JS setup
We use esbuild to compile most of our JavaScript. Jest uses Node to run
our code. This introduces two problems:

- Our code isn't Node-compatible, it's esbuild compatible
- Our code is meant to run in the browser, not in Node

To solve this, we need to bring in two dependencies.

`esbuild-jest` tells Jest to use `esbuild` to precompile any JS/TS(x)
files it finds. We configure it using the `transform` directive in
`package.json`.

`jest-environment-jsdom` configures Jest to run in JSDom, which is a
browser-like Node environment. We configure this using the
`testEnvironment` directive.

The other configuration values in this commit:

- `testPathIgnorePatterns` tells Jest not to look for spec files in
  `node_modules`, saving time
- `transformIgnorePatterns` is set to ignore `node_modules` by default.
  This breaks when we attempt to test code that imports e.g. `workbox`
  which needs to pass through `esbuild` before it can be interpreted by
  Jest. By setting it to `[]`, we tell `esbuild` to not ignore it
  anymore.