Get campaign from URL in offline_campaign_controller
This used to be a hardcoded 1 which messes up tests that have a random
ID for the campaign.

This is a temporary implementation; would probably be better to do this
in a more declarative way, like specifying the ID in the DOM, or simply
not passing it in via payload.