Add store module
This instantiates an indexedDB for storing offline PUT/POST requests in.

The module exposes methods to add, query, and delete items from the
database.

The module is written in TypeScript, and specifies a schema, which means
that errors in respecting the schema should be caught at compile time.

We don't currently include a typescript linting step. We can add this
later.