Add asciidoctor
The Mermaid diagrams are coming out not great. Moving to puml is an easy
fix, however, it does mean we need to generate the images by hand now as
MD dosn't appear to have support to automatically generate them the same
as the mermaid ones.

While we're switching away from having the diagram inline and to
pre-generating them, I'm taking the opportunity to replace Markdown with
asciidoc as it supports generating the plantuml diagram from inline puml
code, plus other useful features such as built-in support for a table of
contents.