Update local development port to 4000
When we develop locally, the serviceworker registers at localhost:port.
If we switch to develop on other apps, the serviceworker stays
registered. This can cause issues.

One way around this is to use a less common port, like 4000.

We could switch to using a custom domain in the future, but that would
be more complex.

Also remove hardcoded localhost from record/child template URLs.