Set the Puma to port 4444 if tmp/offline.txt exists
This is a simple "offline-mode" implementation for Rails. The idea is to
take down the server and restart it on a different port when we want to
switch from offline to online mode and vice versa.

With this config, you can take down the server with a script line:

```
touch tmp/offline.txt && kill -USR2 $(cat tmp/pids/server.pid)
```

And take it back online with:

```
rm -f tmp/offline.txt && kill -USR2 $(cat tmp/pids/server.pid)
```