Add /ping endpoint
/health is more useful for things like blue/green deployments and
testing if the application has successfully deployed, and can be
configured to do more work than simply replying with "200 OK". It could
e.g. check for downstream services like Notify, or perhaps a FHIR API we
depend on for essential functionality.

If all we are checking for is if the user has a working internet
connection and can route successfully to a server we control, then a
simpler `/ping` endpoint is more relevant.

Using a separate endpoint also allows us to silence logging on it.