Replace architecture markdown with asciidoc
Asciidoc provides a better markup language and has built-in support for
plantuml out of the box, which renders much more nicely than mermaid.
However, GitHub doesn't currently support inline rendering of Asciidoc
files, which is a downside, but the tooling around Asciidoc is better
because it supports rendering PlantUML on the CLI. So the reasoning is:

1. PlantUML renders noticeably better than Mermaid.
2. GitHub's Markdown rendering doesn't support PlantUML.
3. So to support PlantUML rendering we would have to pre-process the
   Markdown file. There is possibly an established way to do this with
   MD, HOWEVER ...
4. Asciidoc supports this with a well established asciidoctor extension,
   and with the other benefits asciidoc brings, including in-browser
   rendering support through a browser plugin, this seems to be a better
   choice.

Also, as an addendum, PlantUML and Mermaid syntax is pretty much the
same, so should we decide to migrate back to Markdown+Mermaid it should
not be a lot of extra work.