Render PlantUML diagrams locally
This allows us to render diagrams locally while updating documentation.
Without this we'd have to commit and push changes to see the complete
document while working on it.

The downside to this is that it adds boiler-plate code for each diagram.
I tried to include partials but GitHub renders asciidoc files in safe
mode which disables include. Unfortunately this is the only way to do
this.