Replace link to / with /dashboard in header
Caching redirects for offline use isn't possible using our IndexedDB
approach. When our ServiceWorker's event listener for `fetch` events
receives the request for `/`, which is a redirect, the response object
is of type `opaqueredirect` and contains no information about where the
redirect is meant to go. This is by design: https://github.com/whatwg/fetch/issues/763

Alternatives to handling redirects in the application when offline
include:

- Falling back to the Cache API to handle these types of requests
- Writing some very brittle code that depends on a redirect request +
  its follow-up request to happen sequentially, which sounds hacky
- Simply not using redirects where possible when designing the core
  navigation and features of our app

The latter option is easiest.