Set all pages to Cache-Control no-store
Without this, the browser can choose to temporarily store pages in the
user's navigation in a client-side plaintext cache. Because our pages
can contain things like NHS numbers or user data, this is a potential
security risk.

There's no tests in this commit because `spec/requests/pages_spec.rb`
covered testing the `/children.json` endpoint, and it should still pass.