Add additional pages to cache via addAll
Since we're no longer relying on user navigation to populate the cache,
we need to add additional pages to the SAVE_CAMPAIGN_FOR_OFFLINE event.

The paths to the CSS and JS are fingerprinted, so we need to fish these
out of the frontend, which can be done in the
offline_campaign_controller.