Add wait to offline test
When clicking "Save offline," 5 or so requests are queued from the
server. If we immediately switch the server to offline, those requests
might still be in flight. We could use the `swWaitForRequest` function
with a loop of all the possible requests to wait for them, or we could
just add a 100ms timeout, which seems simpler.

This is a temporary measure to fix flaky tests on the `main` branch and
will be replaced by a different offline e2e test, once we switch to the
password-based implementation.