Add offline password setup pages
This adds the model, view, and controller for the /setup-offline route.

This was split off from #198
so it will be linked to from a follow-up PR.

The OfflinePassword model will be used to validate and save a backup of a user's offline password.

For now, passwords are stored in plain text in the database. In the
future we can decide to do a number of other things:

- Save the derived key instead of the password
- Save the password using ActiveRecord Encryption, a Rails 7 feature, so
  that passwords are not stored in the clear, but are still recoverable