Handle submitting an invalid password
We need to only save assets and data offline if we have a proper offline
password setup.

To do this, we need to pass the password to the server. If the response
from the server is successful, we should save the user's campaign for
offline working.

To do this, we've set up a service worker POST route handler that
intercepts the request. It prevents the default form action and instead
launches its own fetch to the server.

The password is retrieved from the form data and then used to initialise
the encrypted cache. The CSS and JS are passed in as form parameters,
because they can't be retrieved from the service worker as it can't
access the DOM.