Update offline spec to continue requests
The offline spec had a race condition before this change.

It would wait for a /ping request, fulfill it immediately (which doesn't
let it go to the server), and then refresh the page. It just so happened
that most of the time, the refresh would come after the PUT
/sessions/1/vaccinations/record resolve resolved. But not always.

This changes it to always wait for the PUT /record, and to also send it
upstream by switching from .fulfill to .continue.