Create PatientSession model
We'll need to record multiple vaccination outcomes per patient +
session. To model this, we've decided to have a one-to-many relationship
between patient_sessions and vaccination_records.

To create that relationship, Rails wants the PatientSession to be its
own join model, rather than just a join table. We also have to
explicitly add `has_many :patient_sessions` to the patient and session
models.

This change does not affect the controller or view code at all, as its
just plumbing to update the table name and to associate the models
correctly.