Update example data
Since we're using a JSON file for our example campiagn it's important to
update it with any new data we add. This change adds a script that
generates (semi-)random consent data and updates the example data with
it. This keeps our example campaign consistent across environments.

One caveat is that this new rake task relies on a customised version of
Faker, which adds a "alternative_name" generator to create common name
alternatives from a first name. For example "Arnie" is an common
alternative to "Arnold". This should be low impact since Faker doesn't
change often, but the intention will be to try to merge this with Faker
upstream.