Nest triage route under sessions/patients
When it came to adding an update it started to become apparent that
nesting triage directly under sessions was incorrect. Mainly because it
isn't RESTful so figuring out what resource and ids to use started to
become unaparrent and surprising.

This change makes the routes more restful and tidies other bits up in
the triage controller.