Update vaccinations end to end spec
This should have been done as part of the vaccinations show page update,
but it was skipped, because in order to do it, we need to disable the
`fhir_server_integration` flag in the tests. This breaks an existing
request spec. It's been commented out for now, as the `fhir` integration
is less important than the model office work.