Create vaccination_record when vaccinating
This migrates us off of the `patient.seen` field and instead uses a
separate `VaccinationRecord` model to record the event.

The model is initialised on the vaccinations#show page, which POSTs to
the #confirm action. This action attempts to set the administered and
site fields on the model. If it fails, the user has to address the
errors before continuing.

When it succeeds, users arrive at the #confirm view and check their
answers. After checking, the record is saved, administered_at is set to
the current timestamp, and the status tag on the index page is updated.