Add patient session state machine concern
This is a spike to add a patient session state machine via a concern.
It will replace the current code that attempts to derive the state
from the various objects with a relatively low-fidelity state field
which tells us what point in the lifecycle the patient's session record
is in, but without telling reflecting too much of the state.

As an example, there is a state for "awaiting_triage", however this does
not reflect whether the consent has been given or refused, that is
reflected on the consent record and that must be examined to know the
full state.