Refactor patient sessions state names
State names should reflect the condition or state the record is in. For
example, "Awaiting triage" talks about the action that will be taken on
the record, not about what state it's actually in, which is that a
consent response has been received. Displaying a record that has consent
as needing triage should be a part of the logic of how we display this
state to the user.