Refactor state and transition names
This makes it clearer that the state machine is processing the states
and deciding on the outcome of the transition. This means that the
business logic for moving between states is kept in this concern and
callers don't need to replicate the logic.

This change also adds more states to more richly describe what state the
patient is in for that session.