Replace GOVUK Frontend initAll with stimulus shims
initAll is an easy way to initialise all GOVUK Frontend elements on the
page with their custom JS. However, it doesn't allow initialisation of
individual modules, and doesn't work well with Rails Turbo, because the
elements are only initialised once and not on every page navigation.

The main reason to do this refactoring is to allow us to replace the JS
for GOVUK Radios with a shim that's cross-compatible with the NHSUK
Radios.