Create .nhsuk-visually-hidden class
The actual class is .nhsuk-u-visually-hidden, but when we use the
`config-brand` option for the GOVUK Form Builder to configure the
classes, it ends up inventing a .nhsuk-visually-hidden class. While we
could go and monkey patch all the relevant components to use the
correct class, it's easier to just add a new CSS rule.