Replace GOVUK Radios JS with NHSUK Radios JS
This implements a Stimulus controller shim to enhance `nhsuk-radios`
using the custom JS from `nhsuk-frontend`.

Because we're using GOVUK Radios HTML but with NHSUK Radios styling, we
need to manually add a `nhsuk-radios--conditional` class to the root
element.

In addition, NHSUK Radios use `aria-controls` on the server-rendered
HTML, but GOVUK Radios use `data-aria-controls` and enhance it when JS
initialises. Using `data-aria-controls` is the better solution, because
it means that the attribute doesn't kick in in a no-JS situation.