Rename "action" to "status" when displaying state
Our old naming was "action or outcome", which got shortened to just
"action" in a recent change. "status" may be better as, for now at
least, this value is only used in views (we rely on "state" internally),
so differentiating between "action" or "outcome" doesn't have any
value (yet). And also "status" is also shorter than "action_or_outcome".