Add consent journey confirmation page
This is the first step in building the verbal consent journey. We're
starting with the "Check your answers" page which is the penultimate
step, and will get it to work with stubbed data to begin with. Follow-up
PRs will uncomment the rest of the lines in the end to end test and
progressively build out the rest of the functinality.