Reduce expect timeout to 500ms
Playwright likely uses a generous default timeout here to work better
with single page apps that might rely on asynchronous HTTP requests to
fetch in data and update the page. Because we have a multi-page app,
generally speaking, when the page has finished loading, the content
should be there. So this timeout can be reduced.