Render health questions form
This allows users to fill in the health questions when they are
submitting verbal consent notes.

A `.fields_for` call on a locally instantiated `HealthQuestion` model is
used to pre-populate the fields with the values on the
`health_questions` field. This changes the create method to only seed
the question titles, not the answers, which are then subsequently used
to render this list.

This is brittle, crucially it's missing validation and can be easily
broken. It's enough for the model office but should be revisited soon.