Use asdf node version in Jest CI run
We're currently using the latest stable Node.js, which seems to have
changed something TextDecoder related which now causes our Jest tests to
fail. We should be using the one we specify in `.tool-versions`.
`setup-node` does not pick it up automatically, but we can pick it up
manually with `grep` and `awk`.