Add default batch functionality
Allows nurses to select a batch to be used by default for their session.
This gets stored in their session object and expires the next day. When
a default batch has been selected, the batch selection screen is skipped
and they are taken straight to the confirmation page when recording a
vaccination.