Remove consent spec
The feature of a nurse receiving and recording consent while in triage
is already tested as part of the full journey test, so this test isn't
necessary on it's own. Additionally, this test tested recording a
consent refusal followed by consent being given, which may not be a
scenario we need to test (if a nurse has called a parent and they've
refused consent, would they ever call them back to get consent again?).
If we need to support that scenario we can do that as a specific feature
test.