Render error when revisiting old consent forms
Users can use the browser back button to return to old forms that are
part of the consent journey, resulting in undefined behaviour in the
app. This change makes it clearer that what they are doing isn't
supported by rendering errors when form pages are requested in an
inconsistent state.