Define steps in ConsentForm
Replace the `on:` context based validation with conditional validation
that relies on specifying a `form_step`. This also makes the validations
fire off for previous steps, and at the end for the final `record` step.

Unifies the `*_params` methods in the EditController into one
`update_params` method.

`required_for_step?` takes an optional `exact` parameter, for transient
steps like the `school` one that ask a question but don't actually save
a field to the database (and shouldn't be re-validated on other steps).

Overall approach based on:
https://www.joshmcarthur.com/2014/12/23/rails-multistep-forms.html