Make edit controller steps dynamic
Before, `steps` was statically called in the controller class using the
also statically defined `ConsentForm.form_steps` cattr_accessor. This
doesn't work if we want to have a branching wizard with multiple
possible paths depending on the user's choice.

This implements a `form_steps` instance method and then calls a
`:set_steps` before action in the EditController that initialises
`self.steps` after the consent form is fetched.