Add AppFlashMessageComponent
The older component was only designed for the success messages we
display, but didn't handle warnings, alerts, etc. This becomes an issue
when implementing the Devise sign_in page and displaying sign_in errors.

This implementation allows the flash value to be a simple String of the
message, or a Hash that optionally includes a title and a heading.