Add e2e test for session resource to return error
This change also enables rendering error pages in test env. This means
when an error occurs we won't have the error and stack trace visible in
the video Playwright generates, we'll have to look in logs/test.log to
get that. It does, mean however, that we'll be able to test the error
page that gets rendered in the e2e tests.