Implement authentication for consent forms
This stores the `consent_form.id` in a session variable after initial
creation, and checks it on all subsequent requests through the
edit_controller or the record action. If the ID in the URL does not
match the one stored in the session, we redirect the user to the /start
page so they can start a new consent form.

This prevents users from editing other people's consent forms, but
doesn't allow users to edit multiple consent forms at the same time in
multiple tabs, or to easily resume their work should they close their
tab or if their device loses power.