Allow loading additional example campaign
The `load_campaign_example` task used to only find-and-replace the
campaign it was loading, making it impossible to load an additional
campaign. This change leaves that as the default behaviour, but adds an
option to add the campaign being loaded as a new campaign.

This will be used for testing, for example, to load up multiple
campaigns.