Add hints to HealthAnswer model
Because the questions are saved when the ConsentForm is created, we need
to persist them in the jsonb field containing all the health question
metadata. We could change in the future to persisting just an ID and
then fetching the latest information from the database, but this could
backfire if SAIS teams end up reusing the same question and completely
changing it instead of adding a new one. We'd have to be careful about
how we design the "update" action to disallow editing the question after
it's already been used.