Add good_job gem
good_job is a Postgres-based ActiveJob backend for Rails.

Unlike Sidekiq, it doesn't require a separate Redis instance, and can
run in the main web server process instead of a separate worker. This
means that a good_job deployment is easier to think about (1 server
instead of 3) and costs less, which makes a lot of sense for our small
app.