Implement email validator
This is an email validator that's implemented to a similar spec as the
GOVUK Notify one. This reduces the chance that emails entered by our
users will be rejected by Notify.

It follows their implementation and implements their test cases:

- https://github.com/alphagov/notifications-utils/blob/9311851c105237dc5d770c9f1c0a5ad4c2acb380/notifications_utils/recipients.py#L622
- https://github.com/alphagov/notifications-utils/blob/9311851c105237dc5d770c9f1c0a5ad4c2acb380/notifications_utils/__init__.py#L14
- https://github.com/alphagov/notifications-utils/blob/9311851c105237dc5d770c9f1c0a5ad4c2acb380/tests/test_recipient_validation.py#L118

International Domain Names (IDNs) are not currently implemented.

The two test cases with non-UTF-8 characters are commented out.