Replace health_* steps with health_question
This means we no longer have individual steps for each health question,
using a param to track which question we're currently on. While this can
make it less clear what the form journey is, it allows us to reuse more
of the Wicked forms mechanisms. It also aligns a little bit with the
fact that the journey through the health questions is dynamic.