Specify docker build platform
Copilot builds containers locally using the platform of the host client.
So on a Mac, it'll attempt to build an ARM64 docker container. We're
using `linux/x86-64`, so we need to specify this.

Peculiarly, this is mentioned to be the default in the documentation,
but demonstrably needs to be set to build correctly on an M1 Mac in our
testing.