Clean up session[:follow_up_changes_start_page]
When users edit one of their health questions, we need to keep track of
the health question that they started on, so that we can correctly skip
them to the confirmation page when they finish editing a health
question, or when they use the back button.

Once the health question is edited and they've finished, we should clean
this variable up, so that it doesn't interfere with situations where the
user then edits something else, like their consent.

Before this, the session variable was lingering on the session and
causing bugs in a test that submits multiple consent forms in a row.