Replace patient_session#consent with #consents
We are migrating the app to work with multiple consents. This removes
the #consent_for_campaign method on the patient and the #consent method
on the patient session and adds a single #consents method to the latter.

The situations where #consent_for_campaign was being used directly on
the patient can all be replaced with patient_session#consents.