Replace hard-coded names in tests with fixture data
This is current as of the latest generated data. We'll have to update it
whenever the data changes. That shouldn't be too much effort, and we can
automate it in the future.

We decided that this is a simpler approach than trying to click the
"first row" of a certain kind, which would have required significant
re-engineering.