Match mum/dad to patient records
This builds on the previous `match_mum_and_dad_info_to_consent` method.
The previous way would massage the patient data based on the randomly
generated consent data.

This new approach instead purposefully generates a number of consent
objects that plausibly match the patient details. It also makes sure
that situations where there are 2 consents are uniformly from a mum and
a dad, and that situation with 3 consents are from a mum, dad, and
granddad.