Add PatientSession#ready_to_vaccinate?
This will be used by AppVaccinateFormComponent to determine whether to
render the form or not. This differs slightly from
PatientSession#able_to_vaccinate in that it's this new predicate is used
simply to check if the patient is in a state where the nurse should be
prompted to record a vaccination.