Only show triage form when needed
The triage form should only show when the patient_session is in the
appropriate state.

This change adds the concept of a predicate that checks what is the next
step for this patient, "next_step_is_triage?" in this case. This differs
from the guards used in the state machine that check if patient_session
is _permitted_ to perform a transition whereas this new check is to see
if it is expected to do this transition next. If this approach tracks
then we should consider rolling it out to other places such as the
status banners and other parts of the patient page.