Display health questions correctly
If there is only one consent, we should just display the formatted
answer.

If there are two or more consents, but they are matching, we should
display "All responded:" followed by the formatted answer.

If there are two or more consents, but they are conflicting, we should
display "X responded: <answer>" for each consent.