Remove Get consent tab from Triage
This functionality is now handled by the No response tab in the Check
consent responses section. Update the tests accordingly.

It's worth mentioning that the tests currently pass, but that is because
at the moment submitting consent will send the user to the Triage index.
This is a bug, and when we fix it, these tests will have to be updated
again.

This also removes spec/requests/triage_spec.rb, because the
functionality that it was testing for is already covered in the
Playwright tests.