Optimise :consents lookup on patient_session
Instead of using a def consents method, we can instead create a custom
scope on Consent and then fetch it using a has_many association. This
allows us to `includes(:consents)`.

It's worth mentioning that because this is an instance dependent scope,
using it means that we won't be able to combine it with things like
`order_by` in certain scenarios.