Replace update! calls with save(validate: false)
The consent model does not have any validations that run on the base
model without a validation context. However, we're adding some of those
in the new consents journey.

To make this work, we have two choices: either make the new validations
dependent on the feature flag, or replace the old calls to `.update!`
with `save(validate: false)`, which is functionally equivalent to what
they were doing before, as there were no validations on the base
context.