Use random number in SAIS team name not county
Although it makes sense to use the county in the SAIS team's name,
with the way we generate multiple sessions now it makes it more
complicted to use it as part of the team name. Not worth the effort
since we don't see the the team name anywhere currently. Embrace and
increment later.