Do not create patients by default for sessions
Seems like a minor thing, but it speeds up the specs by 10% and I don't
think we should create any objects we don't need in factories. e.g. for
a session we _need_ a campaign, but we don't need a patient object.
Those should be added as required in testing scenarios.