Switch seed used for flu campaign
In changing how mobile phone numbers are generated in the patients
factory, we appear to have fallen down a crevasse where the team names
generated for our example campaigns are identical. This doesn't happen
for every seed, but it seems to be more common when the seeds are the
same.

Having the same team name causes issues when loading the example
campaigns because the teams are created with "find_or_*" methods. This
could be fixable in the campaign loader, but changing the seed used for
the flu campaign seems easier and still quite sensible