Replace foreman with concurrently
`foreman` doesn't properly forward the echo of input to commands. This
means that when using the `debugger` or `binding.pry` statements in
Rails, we can't see anything that's being typed into the REPL. This
makes debugging difficult.

`concurrently` is a JS library that can be used as an alternative to
foreman.

More context: https://github.com/ddollar/foreman/issues/703