Handle consent response changing
If a nurse calls a parent and they change their consent response from
refused to given, we need to re-initialise the health answers.
Similarly, if they change from a response given to a response refused,
we need to delete their health answers.

In both scenarios, we want to reset the `reason_for_refusal`.